Enable opengl support for tuna device but not for cards
This will enable overal opengl support but not for the WebAppManager as this is disabled
due to a specific #ifndef line in Src/webbase/WebAppManager.cpp but it will improve
especially all rendering aspects of the main process like cards management.

OpenGL accelerated rendering inside cards is not support by our EGL/GLES drivers yet so we
can't use it.