Rework wifi network security type handling
With the implementation of the webos-connman-adapter the format of the response for the
findnetworks request has changed. The resonse does not contain the securityType field
anyl longer but as replacement a list of available security types. As the existing
implementation relies on the securityType field these changes rework the handling of the
security information and uses a boolean flag only to indicate wether the network is
secured or not instead of passing the available security types around for nothing.

Handling a specific security type should be up to the application which does the wifi
network management and shows the dialog to the user where he has to enter his credentials.

Signed-off-by: Simon Busch <morphis@gravedo.de>